Fork github action to delete packages
Create a new repo with a github action to find and delete packages
copied mostly from the official github action. Update some code that
looks up the old versions to include a search filter so I can remove
only versions that match the current branch. Update the readme to
reflect all of this.